Added new rooms and details for the main room.
Create a new scene for the box, cirlce, and blank room. the circle room is still a box because Unity does not have a cyclider we can use for the room. I made it so that the user can navigate between the rooms. Press the start button while in the "New" rooms to bring up the "pause" menu.