Fixed right hand only, double models, save name limit, and buttons too big
Pressing either ot the triggers with swtich the land and right mode. If the left trigger is pressed, the program switches to left hand mode. If the right hand is triggered, the right mode is activated. No new model may been downloaded until the current one is finished being placed from download. Save name limit has been set to 20. Only one button is allowed to be highlighted or selected at a time.